Move to proper directories
