+++ Docker/Configuration/Usage/Adapter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Keboola\DockerBundle\Docker\Configuration\Usage;
+
+use Keboola\DockerBundle\Docker\Configuration;
+
+class Adapter extends Configuration\Adapter
+{
+    protected $configClass = Configuration\Usage::class;
+}
